mseek : update after https://codereview.scilab.org/#/c/12551/
/!\ LONG TIME EXECUTION  /!\
exec("/home/ced/work/scilab/scilab/modules/fileio/tests/nonreg_tests/bug_6511.tst")

+++ scilab/modules/fileio/includes/mseek.h
@@ -19,7 +19,7 @@
 * seek function
 */
 
-FILEIO_IMPEXP int mseek(int fd, int offset, int flag);
+FILEIO_IMPEXP int mseek(int fd, long long offset, int flag);
 
 #endif /* __MSEEK_H__ */
 /*--------------------------------------------------------------------------*/
